+++ FS.FilterExpressionCreator.Tests/Tests/TypeFilter/FilterForShortByValueTests.cs
@@ -46,7 +46,10 @@ namespace FS.FilterExpressionCreator.Tests.Tests.TypeFilter
             FilterTestCase.Create(1101, FilterOperator.Default, new short[] { -10 }, (short _) => NONE),
             FilterTestCase.Create(1102, FilterOperator.Default, new short[] { +5 }, (short x) => x == +5),
 
-            FilterTestCase.Create(1200, FilterOperator.Contains, new short[] { 0 }, new FilterExpressionCreationException("Filter operator 'Contains' not allowed for property type 'System.Int16'")),
+            FilterTestCase.Create(1200, FilterOperator.Contains, new short[] { +5 }, (short x) => x == +5 || x == -5),
+            FilterTestCase.Create(1201, FilterOperator.Contains, new short[] { -5 }, (short x) => x == -5),
+            FilterTestCase.Create(1202, FilterOperator.Contains, new short[] { +3 }, (short _) => NONE),
+            FilterTestCase.Create(1203, FilterOperator.Contains, new short[] { +0 }, (short x) => x == 0),
 
             FilterTestCase.Create(1300, FilterOperator.EqualCaseInsensitive, new short[] { -5 }, (short x) => x == -5),
             FilterTestCase.Create(1301, FilterOperator.EqualCaseInsensitive, new short[] { -10 }, (short _) => NONE),